refactor code

+++ ProcessingCSV/src/main/java/org/example/validationChain/SalaryValidator.java
@@ -6,7 +6,7 @@ public class SalaryValidator extends BaseValidator {
     @Override
     public boolean validate(String[] data) {
         try {
-            int salary = Integer.parseInt(data[3].trim());
+            float salary = Float.parseFloat(data[3].trim());
             return salary >= 0 && super.validate(data);
         } catch (NumberFormatException e) {
             return false;
